feat: 3개의 창 컴포넌트로 분리
