+++ src/pages/AssignmentPage/ProblemSolvePage.css
@@ -230,349 +230,6 @@
   min-height: 0;
 }
 
-/* Description Area */
-.description-area {
-  padding: 0 0 24px 0;
-  overflow: auto;
-  background: linear-gradient(135deg, #0d1117 0%, #161b22 100%);
-  border-right: 1px solid #30363d;
-  box-shadow: inset 0 0 10px rgba(0,0,0,0.3);
-}
-
-.problem-solve-page.light .description-area {
-  background: linear-gradient(135deg, #f8f9fa 0%, #ffffff 100%);
-  border-right: 1px solid #e1e4e8;
-  box-shadow: inset 0 0 10px rgba(0,0,0,0.05);
-}
-
-.description-area > div {
-  padding: 0 24px;
-}
-
-.description-header {
-  margin: 0 0 16px 0;
-  font-size: 16px;
-  font-weight: bold;
-  padding: 12px 16px;
-  color: #58a6ff;
-  background: linear-gradient(90deg, rgba(88, 166, 255, 0.1) 0%, rgba(88, 166, 255, 0.05) 100%);
-  border-bottom: 1px solid #30363d;
-  border-radius: 6px 6px 0 0;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  flex-wrap: wrap;
-  gap: 12px;
-  min-height: 48px;
-  box-sizing: border-box;
-}
-
-.problem-solve-page.light .description-header {
-  color: #0969da;
-  background: linear-gradient(90deg, rgba(9, 105, 218, 0.1) 0%, rgba(9, 105, 218, 0.05) 100%);
-  border-bottom: 1px solid #e1e4e8;
-}
-
-/* Problem Limits in Header */
-.problem-limits-header {
-  display: flex;
-  gap: 8px;
-  flex-wrap: wrap;
-}
-
-.limit-badge-header {
-  font-size: 12px;
-  font-weight: 500;
-  padding: 4px 8px;
-  border-radius: 4px;
-  display: inline-flex;
-  align-items: center;
-  white-space: nowrap;
-}
-
-.limit-badge-header.time-limit {
-  background: rgba(255, 255, 255, 0.15);
-  color: #58a6ff;
-  border: 1px solid rgba(88, 166, 255, 0.3);
-}
-
-.limit-badge-header.memory-limit {
-  background: rgba(255, 255, 255, 0.15);
-  color: #ff6b6b;
-  border: 1px solid rgba(255, 107, 107, 0.3);
-}
-
-/* Light theme for header badges */
-.problem-solve-page.light .limit-badge-header.time-limit {
-  background: rgba(9, 105, 218, 0.15);
-  color: #0969da;
-  border: 1px solid rgba(9, 105, 218, 0.3);
-}
-
-.problem-solve-page.light .limit-badge-header.memory-limit {
-  background: rgba(220, 53, 132, 0.15);
-  color: #d63384;
-  border: 1px solid rgba(220, 53, 132, 0.3);
-}
-
-/* Description Content */
-.description-content {
-  padding: 16px 0;
-  line-height: 1.6;
-  color: #EFF5F2;
-}
-
-.problem-solve-page.light .description-content {
-  color: #000000;
-}
-
-.description-content h1,
-.description-content h2,
-.description-content h3,
-.description-content h4,
-.description-content h5,
-.description-content h6 {
-  margin: 16px 0 8px 0;
-  color: #EFF5F2 !important;
-  font-weight: bold;
-}
-
-.problem-solve-page.light .description-content h1,
-.problem-solve-page.light .description-content h2,
-.problem-solve-page.light .description-content h3,
-.problem-solve-page.light .description-content h4,
-.problem-solve-page.light .description-content h5,
-.problem-solve-page.light .description-content h6 {
-  color: #000000 !important;
-}
-
-.description-content p {
-  margin: 8px 0;
-}
-
-.description-content ul,
-.description-content ol {
-  margin: 8px 0;
-  padding-left: 24px;
-}
-
-.description-content li {
-  margin: 4px 0;
-}
-
-.description-content code {
-  background-color: #30363d;
-  color: #e1e4e8;
-  padding: 2px 6px;
-  border-radius: 3px;
-  font-family: 'Monaco', 'Menlo', 'Ubuntu Mono', monospace;
-  font-size: 0.9em;
-}
-
-.problem-solve-page.light .description-content code {
-  background-color: #f6f8fa;
-  color: #24292e;
-}
-
-.problem-solve-page.light .due-date-info {
-  background: rgba(255, 107, 107, 0.1);
-  border-left-color: #ff6b6b;
-}
-
-.problem-solve-page.light .due-date-content {
-  color: #333333;
-}
-
-.problem-solve-page.light .due-date-content.overdue {
-  color: #d63384;
-}
-
-.description-content pre {
-  background-color: #0d1117;
-}
-
-/* Problem Limits Info */
-.problem-limits-info {
-  margin: 16px 0;
-  padding: 16px;
-  background: rgba(88, 166, 255, 0.1);
-  border-radius: 8px;
-  border-left: 4px solid #58a6ff;
-}
-
-.limits-header {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-weight: 600;
-  font-size: 14px;
-  color: #58a6ff;
-  margin-bottom: 12px;
-}
-
-.limits-icon {
-  font-size: 16px;
-}
-
-.limits-content {
-  display: flex;
-  flex-direction: column;
-  gap: 8px;
-}
-
-.limit-item {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-}
-
-.limit-label {
-  font-size: 14px;
-  color: #e0e0e0;
-  font-weight: 500;
-}
-
-.limit-value {
-  font-size: 14px;
-  color: #58a6ff;
-  font-weight: 600;
-  background: rgba(88, 166, 255, 0.15);
-  padding: 4px 8px;
-  border-radius: 4px;
-}
-
-/* Light theme for problem limits */
-.problem-solve-page.light .problem-limits-info {
-  background: rgba(9, 105, 218, 0.1);
-  border-left-color: #0969da;
-}
-
-.problem-solve-page.light .limits-header {
-  color: #0969da;
-}
-
-.problem-solve-page.light .limit-label {
-  color: #333333;
-}
-
-.problem-solve-page.light .limit-value {
-  color: #0969da;
-  background: rgba(9, 105, 218, 0.15);
-}
-
-/* Due Date Info */
-.due-date-info {
-  margin: 16px 0;
-  padding: 16px;
-  background: rgba(255, 107, 107, 0.1);
-  border-radius: 8px;
-  border-left: 4px solid #ff6b6b;
-}
-
-.due-date-header {
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  font-weight: 600;
-  font-size: 14px;
-  color: #ff6b6b;
-  margin-bottom: 8px;
-}
-
-.due-date-icon {
-  font-size: 16px;
-}
-
-.due-date-content {
-  font-size: 14px;
-  color: #e0e0e0;
-}
-
-.due-date-content.overdue {
-  color: #ff6b6b;
-  font-weight: 600;
-}
-
-.overdue-warning {
-  color: #ff6b6b;
-  font-weight: 600;
-  margin-left: 8px;
-}
-
-.description-content pre {
-  background-color: #0d1117;
-  color: #e1e4e8;
-  padding: 16px;
-  border-radius: 6px;
-  overflow-x: auto;
-  margin: 12px 0;
-  border: 1px solid #30363d;
-}
-
-.problem-solve-page.light .description-content pre {
-  background-color: #f6f8fa;
-  color: #24292e;
-  border: 1px solid #e1e4e8;
-}
-
-.description-content pre code {
-  background-color: transparent;
-  padding: 0;
-  border-radius: 0;
-}
-
-.description-content blockquote {
-  border-left: 4px solid #58a6ff;
-  margin: 12px 0;
-  padding: 8px 16px;
-  background-color: rgba(88, 166, 255, 0.1);
-  border-radius: 0 4px 4px 0;
-}
-
-.problem-solve-page.light .description-content blockquote {
-  border-left-color: #0969da;
-  background-color: rgba(9, 105, 218, 0.1);
-}
-
-.description-content table {
-  border-collapse: collapse;
-  width: 100%;
-  margin: 12px 0;
-}
-
-.description-content th,
-.description-content td {
-  border: 1px solid #30363d;
-  padding: 8px 12px;
-  text-align: left;
-}
-
-.problem-solve-page.light .description-content th,
-.problem-solve-page.light .description-content td {
-  border: 1px solid #e1e4e8;
-}
-
-.description-content th {
-  background-color: #21262d;
-  font-weight: bold;
-}
-
-.problem-solve-page.light .description-content th {
-  background-color: #f6f8fa;
-}
-
-.description-content a {
-  color: #58a6ff;
-  text-decoration: none;
-}
-
-.problem-solve-page.light .description-content a {
-  color: #0969da;
-}
-
-.description-content a:hover {
-  text-decoration: underline;
-}
 
 /* Editor Area */
 .editor-wrapper {
